+++ lib/src/layout/remove_item.dart
@@ -1,7 +1,7 @@
 import 'package:docking/src/layout/docking_layout.dart';
 import 'package:docking/src/layout/layout_modifier.dart';
 
-/// Removes a [DockingItem] from this layout.
+/// Removes [DockingItem] from this layout.
 class RemoveItem extends LayoutModifier {
   RemoveItem({required this.itemToRemove});
 
@@ -44,7 +44,9 @@ class RemoveItem extends LayoutModifier {
       if (children.length == 1) {
         return children.first;
       }
-      return DockingTabs(children);
+      DockingTabs newDockingTabs = DockingTabs(children);
+      newDockingTabs.selectedIndex = dockingTabs.selectedIndex;
+      return newDockingTabs;
     } else if (area is DockingParentArea) {
       List<DockingArea> children = [];
       area.forEach((child) {
@@ -59,9 +61,13 @@ class RemoveItem extends LayoutModifier {
         return children.first;
       }
       if (area is DockingRow) {
-        return DockingRow(children);
+        DockingRow row = DockingRow(children);
+        row.weights = area.weights;
+        return row;
       } else if (area is DockingColumn) {
-        return DockingColumn(children);
+        DockingColumn column = DockingColumn(children);
+        column.weights = area.weights;
+        return column;
       }
       throw ArgumentError(
           'DockingArea class not recognized: ' + area.runtimeType.toString());