+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.4.0]
 
 * Method to add item
+* Keep the row and column children's weight, and the selected tab's index after rearranging the layout
 
 ## [0.3.0]